Update rae-ubuntu-install.sh - Adding noninterative option with env var 'DEBIAN_FRONTEND' - Removing innecesaries comments

+++ scripts/rae-ubuntu-install.sh
@@ -18,6 +18,12 @@ function msg_error() {
 }
 
 function create_window() {
+	local LINES=$(tput lines)
+
+	if [ "$DEBIAN_FRONTEND" == 'noninteractive' ]; then
+		return 0
+	fi
+
 	# Create a new window
 	tput smcup
 
@@ -39,6 +45,10 @@ function create_window() {
 function close_window() {
 	local LINES=$(tput lines)
 
+	if [ "$DEBIAN_FRONTEND" == 'noninteractive' ]; then
+		return 0
+	fi
+	
 	# Delete window
 	tput rmcup
 
@@ -55,6 +65,10 @@ function close_window() {
 function prompt_yes_or_no() {
 	local -r PROMPT="$1"
 	local -l answer
+
+	if [ "$DEBIAN_FRONTEND" == 'noninteractive' ]; then
+		return 0
+	fi
 
 	while true; do
 		read -p "$PROMPT [y/n] " answer
@@ -77,8 +91,15 @@ function prompt_yes_or_no() {
 		esac
 	done
 }
+
 function close_graphical_session() {
 	msg_warn 'is required to close the graphical session.'
+
+	if [ "$DEBIAN_FRONTEND" == 'noninteractive' ]; then
+		msg_warn 'Please close the graphical session manually.'
+		return 0
+	fi
+
 	tput sc
 	if prompt_yes_or_no 'Do you want to close it now?'; then
 		loginctl terminate-session $XDG_SESSION_ID || pkill -u $SUDO_USER
@@ -94,7 +115,11 @@ function press_any_key() {
 	tput setab 1
 	# Invisible cursor
 	tput civis
-	read -n 1 -s -r -p ' Press any key to continue...'
+
+	if [ "$DEBIAN_FRONTEND" != 'noninteractive' ]; then
+		read -n 1 -s -r -p 'Press any key to continue...'
+	fi
+
 	# Reset foreground
 	tput sgr0
 	# Reset background
@@ -472,10 +497,6 @@ function arguments_parse() {
 	if [ $# -eq 0 ]; then
 		show_help
 	fi
-
-	#options=$(getopt -o hagdvit --long help,all,install-docker,install-gns3,install-virtualbox,import-docker-image,import-gns3-templates --name "$0" -- "$@")
-
-	#eval set -- "$options"
 
 	while [ $# -gt 0 ]; do
 		case "$1" in
